requirements.txt
numpy~=1.22.3
pandas~=1.3.5
transformers>=4.17.0
tqdm~=4.63.0
torch>=1.10.0
scipy~=1.6.0
scikit-learn~=0.24.0
PyYAML~=6.0
requests~=2.27.1